requirements.txt
flask
pydub
requests
python-magic
markdown
ffmpeg-python
peewee
wavio
numpy